Angles are now doubles as well! \o/

+++ Project/core/src/main/java/dhbw/karlsruhe/it/solar/core/ai/movement/Kinematic.java
@@ -122,7 +122,7 @@ public class Kinematic {
         return rotation;
     }
 
-    public void setRotation(float value) {
+    public void setRotation(double value) {
         this.rotation = value;
     }
